Differential expression between pair of assays

Perform differential expression between a pair of assays using linear (mixed) models

dreamletCompareClusters(
  pb,
  assays,
  method = c("random", "fixed", "none"),
  formula = ~1,
  min.cells = 10,
  useCountsWeights = TRUE,
  isCounts = TRUE,
  normalize.method = "TMM",
  min.count = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)

Arguments

pb
pseudobulk data as SingleCellExperiment object

assays
array of two entries specifying assays (i.e. cell clusters) to compare, of a list of two sets of assays.

method
account for repeated measures from donors using a "random" effect, a "fixed" effect, or "none"

formula
covariates to include in the analysis.

min.cells
minimum number of observed cells for a sample to be included in the analysis

useCountsWeights
use cell count weights

isCounts
logical, indicating if data is raw counts

normalize.method
normalization method to be used by calcNormFactors

min.count
minimum number of reads for a gene to be consider expressed in a sample. Passed to edgeR::filterByExpr

robust
logical, use eBayes method that is robust to outlier genes

quiet
show messages

BPPARAM
parameters for parallel evaluation

...
other arguments passed to dream

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
  assay = "counts",    
  cluster_id = 'cluster_id', 
  sample_id = 'sample_id',
  verbose=FALSE)

# compare first two assays (i.e. cell types)
ct.pairs =  c("B cells", "CD14+ Monocytes")

# run comparison
# use method = 'fixed' here since it is faster
res = dreamletCompareClusters( pb, ct.pairs, method="fixed")

# genes with highest logFC are most highly expressed in 
# B cells compared to CD14+ Monocytes
head(res)

#>           logFC   AveExpr        t      P.Value    adj.P.Val         B
#> CD37   2.868496  8.818581 14.34454 1.530696e-08 8.808458e-07 10.257964
#> PTMA   2.470558 11.235493 13.55444 2.791370e-08 1.262098e-06  9.605840
#> RPSA   2.733516  9.592327 13.27961 3.465659e-08 1.374371e-06  9.411781
#> RPS5   2.460970 10.801916 12.83588 4.956974e-08 1.743202e-06  9.019371
#> EIF4A2 2.757285  7.930468 12.08817 9.293394e-08 2.941359e-06  8.461450
#> RPS27  2.116130 11.095120 11.64721 1.368121e-07 3.951269e-06  7.961504

# compare B cells versus the rest of the cell types
# 'rest' is a keyword indicator all other assays  
res = dreamletCompareClusters( pb, c("B cells", 'rest'), method="fixed")

# genes with highest logFC are most highly expressed in 
# B cells compared to all others
head(res)

#>                             logFC   AveExpr        t      P.Value    adj.P.Val
#> HLA-DRA_ENSG00000204287 2.7280902 10.357105 14.79810 7.842604e-13 7.353463e-11
#> CD74                    2.6457081 11.137156 14.79087 7.919114e-13 7.353463e-11
#> CD37                    2.0697451  8.471062 13.88351 2.764700e-12 2.246319e-10
#> RPS5                    1.1838012 10.964051 12.22845 3.233798e-11 1.910881e-09
#> HERPUD1                 2.1599823  8.737235 11.43198 1.157861e-10 5.017397e-09
#> RPL18A                  0.9783164 11.963469 10.34409 7.366825e-10 2.660243e-08
#>                                B
#> HLA-DRA_ENSG00000204287 19.43714
#> CD74                    19.36458
#> CD37                    18.24811
#> RPS5                    15.56219
#> HERPUD1                 14.46095
#> RPL18A                  12.27541

lst = list( test = c("CD14+ Monocytes", "FCGR3A+ Monocytes"), 
    baseline = c("CD4 T cells", "CD8 T cells"))

# compare 2 monocyte clusters to two T cell clusters
res = dreamletCompareClusters( pb, lst, method="fixed")

# genes with highest logFC are most highly expressed in 
# monocytes compared to T cells
head(res)

#>            logFC   AveExpr        t      P.Value    adj.P.Val        B
#> FTL     4.990465 13.531244 47.94693 1.990977e-19 1.085083e-16 34.81184
#> PSAP    3.999305  9.387747 34.69227 4.360136e-17 1.188137e-14 29.30659
#> TIMP1   6.246062 11.155536 29.56940 6.144924e-16 1.116328e-13 26.83808
#> ANXA5   3.853168  9.699822 28.58047 1.077359e-15 1.467901e-13 26.26882
#> S100A11 3.109519 11.225989 25.10643 9.084420e-15 9.902018e-13 24.16311
#> LGALS3  4.227457  9.664401 23.97963 1.927655e-14 1.750953e-12 23.43323