Differential expression between pair of assays

Perform differential expression between a pair of assays using linear (mixed) models

dreamletPairs(
  pb,
  assay.pairs,
  method = c("random", "fixed", "none"),
  formula = ~1,
  min.cells = 10,
  useCountsWeights = TRUE,
  isCounts = TRUE,
  normalize.method = "TMM",
  min.count = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)

Arguments

pb
pseudobulk data as SingleCellExperiment object

assay.pairs
array of two entries specifying assays (i.e. cell clusters) to compare

method
account for repeated measures from donors using a "random" effect, a "fixed" effect, or "none"

formula
covariates to include in the analysis.

min.cells
minimum number of observed cells for a sample to be included in the analysis

useCountsWeights
use cell count weights

isCounts
logical, indicating if data is raw counts

normalize.method
normalization method to be used by calcNormFactors

min.count
minimum number of reads for a gene to be consider expressed in a sample. Passed to edgeR::filterByExpr

robust
logical, use eBayes method that is robust to outlier genes

quiet
show messages

BPPARAM
parameters for parallel evaluation

...
other arguments passed to dream

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
  assay = "counts",    
  cluster_id = 'cluster_id', 
  sample_id = 'sample_id',
  verbose=FALSE)

# compare first two assays (i.e. cell types)
ct.pairs = assayNames(pb)[1:2]

# run comparison
# use method = 'fixed' here since it is faster
res = dreamletPairs( pb, ct.pairs, method="fixed")

head(res)

#>            logFC   AveExpr        t      P.Value    adj.P.Val        B
#> FTL     4.564199 14.265849 22.83327 1.003928e-10 3.700360e-08 15.23627
#> LGALS3  4.816806  9.879829 22.51767 1.169150e-10 3.700360e-08 14.92328
#> S100A10 4.554573  9.950588 19.90291 4.497204e-10 7.817596e-08 13.68451
#> S100A11 4.691552 10.362616 19.73185 4.940029e-10 7.817596e-08 13.61390
#> CD63    3.888883 10.337664 18.58836 9.447307e-10 1.196029e-07 13.00621
#> ANXA5   3.673970 10.097965 18.04172 1.305381e-09 1.377177e-07 12.68842