Compute logit ratios and their precisions

Compute logit ratios and their precisions given a matrix of counts

Usage

logitRatios(data, pc = 0.25)

# S4 method for matrix
logitRatios(data, pc = 0.25)

# S4 method for SingleCellExperiment
logitRatios(data, pc = 0.25)

Arguments

data

matrix of counts where rows are samples and columns are categories, or object that this information can be extracted from

pc

pseudocount

Value

EList defined by limma storing logit ratio in E and precisions in weights

Details

Let \(c1 \sim \Gamma(\alpha, 1)\), \(c2 \sim \Gamma(\beta, 1)\) and \(x = \frac{c1}{c2}\), then \(x \sim B(\alpha, \beta)\), the expected value of \(x\) is \(E[x] = \frac{\alpha}{\beta}\). The expected log is \(E[\log(x)] = \psi(\alpha) - \psi(\alpha - \beta)\), where \(\psi()\) is the digamma() function.

The variance of the logit transformed fraction is \(var[log(\frac{x}{1-x})] = \psi_1(\alpha) + \psi_1(\beta)\), where \(\psi_1()\) is the trigamma() function. For sufficiently large values of \(\alpha\) and \(\beta\), this is well approximated by \(1/\alpha + 1/\beta\).

The counts producing these ratios are discrete. But approximating the counts by \(\Gamma(\lambda, 1)\) instead of \(Pois(\lambda)\) matches the first two moments exactly.

See https://en.wikipedia.org/wiki/Beta_distribution#Moments_of_logarithmically_transformed_random_variables

Since log() and trigamma() are not defined at zero, add a pseudocount of 0.25.