Run mash analysis on dreamlet results

Run mash analysis on dreamlet results

Usage

run_mash(fit, coefList)

Arguments

fit

result from dreamlet()

coefList

coefficient to be analyzed

Value

a list storing the mashr model as model and the original coefficients as logFC.original

Details

Apply mashr analysis (Urbut et al. 2019) on the joint set of coefficients for each gene and cell type. mashr is a Bayesian statistical method that borrows strength across tests (i.e. genes and cell types) by learning the distribution of non-zero effects based the obesrved logFC and standard errors. The method then estimates the posterior distributions of each coefficient based on the observed value and the genome-wide emprical distribution.

mashr has been previously applied to differential expression in GTEx data using multiple tissues from the same set of donors (Oliva et al. 2020) .

In single cell data, a given gene is often not sufficiently expressed in all cell types. So it is not evaluated in a subsets of cell types, and its coefficient value is NA. Since mashr assumes coefficients and standard errors for every gene and cell type pair, entries with these missing values are set to have coef = 0, and se = 1e6. The output of mashr is then modified to set the corresponding values to NA, to avoid nonsensical results downstream.

References

Oliva M, Munoz-Aguirre M, Kim-Hellmuth S, Wucher V, Gewirtz AD, Cotter DJ, Parsana P, Kasela S, Balliu B, Vinuela A, others (2020). “The impact of sex on gene expression across human tissues.” Science, 369(6509), eaba3066. https://doi.org/10.1126/science.aba3066.

Urbut SM, Wang G, Carbonetto P, Stephens M (2019). “Flexible statistical methods for estimating and testing effects in genomic studies with multiple conditions.” Nature genetics, 51(1), 187--195. https://doi.org/10.1038/s41588-018-0268-8.

See also

Examples

library(muscat)
library(mashr)
#> Loading required package: ashr
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
  assay = "counts",    
  cluster_id = 'cluster_id', 
  sample_id = 'sample_id',
  verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)
#>   B cells...
#> 0.023 secs
#>   CD14+ Monocytes...
#> 0.022 secs
#>   CD4 T cells...
#> 0.023 secs
#>   CD8 T cells...
#> 0.02 secs
#>   FCGR3A+ Monocytes...
#> 0.023 secs

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)
#>   B cells...
#> 0.044 secs
#>   CD14+ Monocytes...
#> 0.067 secs
#>   CD4 T cells...
#> 0.048 secs
#>   CD8 T cells...
#> 0.03 secs
#>   FCGR3A+ Monocytes...
#> 0.062 secs
#> 

# run MASH model
# This can take 10s of minutes on real data
# This small datasets should take ~30s
res_mash = run_mash( res.dl, 'group_idstim')
#> Error in data.frame(assay = k, tab): arguments imply differing number of rows: 1, 0

# extract statistics from mashr model
# NA values indicate genes not sufficiently expressed
# in a given cell type

# original logFC
head(res_mash$logFC.original)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'res_mash' not found

# posterior mean for logFC
head(get_pm(res_mash$model))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'res_mash' not found

# how many gene-by-celltype tests are significant
# i.e.  if a gene is significant in 2 celltypes, it is counted twice
table(get_lfsr(res_mash$model) < 0.05, useNA="ifany")
#> Error in get_lfsr(res_mash$model): object 'res_mash' not found

# how many genes are significant in at least one cell type
table( apply(get_lfsr(res_mash$model), 1, min, na.rm=TRUE) < 0.05)
#> Error in get_lfsr(res_mash$model): object 'res_mash' not found

# how many genes are significant in each cell type
apply(get_lfsr(res_mash$model), 2, function(x) sum(x < 0.05, na.rm=TRUE))
#> Error in get_lfsr(res_mash$model): object 'res_mash' not found

# examine top set of genes
# which genes are significant in at least 1 cell type
sort(names(get_significant_results(res_mash$model)))[1:10]
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'sort': object 'res_mash' not found

# Lets examine CXCL3
# There is a lot of variation in the raw logFC
res_mash$logFC.original["ISG20",]
#> Error in eval(expr, envir, enclos): object 'res_mash' not found

# posterior mean after borrowing across cell type and genes
get_pm(res_mash$model)["ISG20",]
#> Error in get_pm(res_mash$model): object 'res_mash' not found

# gene set analysis using mashr results
library(zenith)
go.gs = get_GeneOntology("CC", to="SYMBOL")
#> 
#> 
# valid values for statistic: 
# "tstatistic", "abs(tstatistic)", "logFC", "abs(logFC)"
df_gs = zenith_gsa(res_mash, go.gs)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'fit' in selecting a method for function 'zenith_gsa': object 'res_mash' not found

# Heatmap of results
plotZenithResults(df_gs, 2, 1)
#> Error in is.data.frame(x): object 'df_gs' not found

# forest plot based on mashr results
plotForest(res_mash, "ISG20")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plotForest': object 'res_mash' not found

# volcano plot based on mashr results
# yaxis uses local false sign rate (lfsr)
plotVolcano(res_mash)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plotVolcano': object 'res_mash' not found