Aggregation of single-cell to pseudobulk data

Aggregation of single-cell to pseudobulk data. Adapted from muscat::aggregateData and has same syntax and results. But can be much faster in for SingleCellExperiment backed by H5AD files

aggregateToPseudoBulk(
  x,
  assay = NULL,
  sample_id = NULL,
  cluster_id = NULL,
  fun = c("sum", "mean", "median", "prop.detected", "num.detected"),
  scale = FALSE,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)

Arguments

x
a SingleCellExperiment.

assay
character string specifying the assay slot to use as input data. Defaults to the 1st available (assayNames(x)[1]).

sample_id
character string specifying which variable to use as sample id

cluster_id
character string specifying which variable to use as cluster id

fun
a character string. Specifies the function to use as summary statistic. Passed to summarizeAssayByGroup2.

scale
logical. Should pseudo-bulks be scaled with the effective library size & multiplied by 1M?

verbose
logical. Should information on progress be reported?

BPPARAM
a BiocParallelParam object specifying how aggregation should be parallelized.

Value

a SingleCellExperiment.

Aggregation parameters (assay, by, fun, scaled) are stored in metadata()$agg_pars, where by = c(cluster_id, sample_id). The number of cells that were aggregated are accessible in int_colData()$n_cells.

Details

Adapted from muscat::aggregateData and has simular syntax and same results. But can be much faster in for SingleCellExperiment backed by H5AD files because this summarizes counts using DelayedMatrixStats.

References

Crowell, HL, Soneson, C, Germain, P-L, Calini, D, Collin, L, Raposo, C, Malhotra, D & Robinson, MD: On the discovery of population-specific state transitions from multi-sample multi-condition single-cell RNA sequencing data. bioRxiv 713412 (2018). doi: https://doi.org/10.1101/713412

Author

Gabriel Hoffman, Helena L Crowell & Mark D Robinson

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# pseudobulk data from each cell type
# is stored as its own assay 
pb

#> class: SingleCellExperiment 
#> dim: 1267 4 
#> metadata(2): experiment_info agg_pars
#> assays(5): B cells CD14+ Monocytes CD4 T cells CD8 T cells FCGR3A+
#>   Monocytes
#> rownames(1267): HES4 ISG15 ... CSTB PRMT2
#> rowData names(2): ENSEMBL SYMBOL
#> colnames(4): ctrl101 ctrl107 stim101 stim107
#> colData names(1): group_id
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):

# aggregate by cluster only, 
# collapsing all samples into the same pseudobulk
pb2 <- aggregateToPseudoBulk(example_sce, cluster_id = "cluster_id", verbose=FALSE)

pb2  

#> class: SingleCellExperiment 
#> dim: 1267 5 
#> metadata(2): experiment_info agg_pars
#> assays(1): ''
#> rownames(1267): HES4 ISG15 ... CSTB PRMT2
#> rowData names(2): ENSEMBL SYMBOL
#> colnames(5): B cells CD14+ Monocytes CD4 T cells CD8 T cells FCGR3A+
#>   Monocytes
#> colData names(0):
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):