Aggregation of single-cell to pseudobulk data

Aggregation of single-cell to pseudobulk data. Adapted from muscat::aggregateData and has same syntax and results. But can be much faster in for SingleCellExperiment backed by H5AD files

aggregateToPseudoBulk(
  x,
  assay = NULL,
  by = c("cluster_id", "sample_id"),
  fun = c("sum", "mean", "median", "prop.detected", "num.detected"),
  scale = FALSE,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)

Arguments

x
a SingleCellExperiment.

assay
character string specifying the assay slot to use as input data. Defaults to the 1st available (assayNames(x)[1]).

by
character vector specifying which colData(x) columns to summarize by (at most 2!).

fun
a character string. Specifies the function to use as summary statistic. Passed to summarizeAssayByGroup2.

scale
logical. Should pseudo-bulks be scaled with the effective library size & multiplied by 1M?

verbose
logical. Should information on progress be reported?

BPPARAM
a BiocParallelParam object specifying how aggregation should be parallelized.

Value

a SingleCellExperiment.

If length(by) == 2, each sheet (assay) contains pseudobulks for each of by[1], e.g., for each cluster when by = "cluster_id". Rows correspond to genes, columns to by[2], e.g., samples when by = "sample_id".

If length(by) == 1, the returned SCE will contain only a single assay with rows = genes and colums = by.

Aggregation parameters (assay, by, fun, scaled) are stored in metadata()$agg_pars, and the number of cells that were aggregated are accessible in int_colData()$n_cells.

Details

Adapted from muscat::aggregateData and has same syntax and results. But can be much faster in for SingleCellExperiment backed by H5AD filesm because this summarized counts using DelayedMatrixStats

References

Crowell, HL, Soneson, C, Germain, P-L, Calini, D, Collin, L, Raposo, C, Malhotra, D & Robinson, MD: On the discovery of population-specific state transitions from multi-sample multi-condition single-cell RNA sequencing data. bioRxiv 713412 (2018). doi: https://doi.org/10.1101/713412

Author

Gabriel Hoffman, Helena L Crowell & Mark D Robinson

Examples

library(muscat)

# pseudobulk counts by cluster-sample
data(example_sce)
pb <- aggregateToPseudoBulk(example_sce, verbose=FALSE)

library(SingleCellExperiment)

#> Loading required package: SummarizedExperiment

#> Loading required package: MatrixGenerics

#> Loading required package: matrixStats

#> 
#> Attaching package: ‘MatrixGenerics’

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars

#> Loading required package: GenomicRanges

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following object is masked from ‘package:limma’:
#> 
#>     plotMA

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following objects are masked from ‘package:base’:
#> 
#>     expand.grid, I, unname

#> Loading required package: IRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> 
#> Attaching package: ‘Biobase’

#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians

#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians

assayNames(example_sce)  # one sheet per cluster

#> [1] "counts"    "logcounts"

head(assay(example_sce)) # n_genes x n_samples

#> 6 x 1556 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 1556 column names ‘ATCATGCTGCGTAT-1’, ‘GTACGAACTGTCAG-1’, ‘CGTGATGAATTTCC-1’ ... ]]

#>                                                                               
#> HES4     . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    2 1 . . . . . . . . 1 . . . . . 2 . . . . . . . . . . . . . . . . . .
#> AURKAIP1 1 . 1 1 . . . . 1 1 . . . . . . . . . 1 . 1 . 2 . . . . . . . . . 1 .
#> MRPL20   . . . . . . 1 1 . . . . . . . . . . 1 . . . . . . . . . . . . . . 1 .
#> SSU72    . . . . . . 1 . . 1 . 1 . . . . . . . 2 1 . . . . . 1 . . 1 . 1 . . .
#> RER1     . . 1 . . . . . 1 . . . . . . . . . . 2 . . . . 1 . 1 . . 1 . 2 2 1 .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . . . . . . . . . . . 1 1 . . . . . . . . . . . . . . 1 . . . .
#> AURKAIP1 . . . . . . . . . . . . . . 1 . 1 . . . 1 . . . . . . . . . . . . 1 .
#> MRPL20   . . . . . . . . . . 1 . . . . . . . . . 1 . . . . . . . 1 . . . . . 2
#> SSU72    . . . 1 1 . 1 1 . . . 1 . . . 1 . . 1 . . 1 . . . . 1 . 1 . . . 1 . 1
#> RER1     1 . . . . . . . . . . . . . . . 1 . . 1 . . . . . . . 1 1 . . . 1 5 .
#>                                                                                
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .  . . .
#> ISG15    . . . . . . . . 1 . . . . . . . . . . . . . . . . 2 . . . . . 1 10 . .
#> AURKAIP1 . . 1 . . . . . 1 . 1 . . . 1 . . 2 . . . . . . . . . . . 1 . 2  . 1 1
#> MRPL20   . . . . . . . . . . . . . . . . . 1 . . . . . . 1 . . . 1 . . 3  . . .
#> SSU72    . . . . . . 1 . 1 . 1 . 1 . . . . . 1 . . . . . . 1 . . . . . 1  . 1 .
#> RER1     . . . 2 . . . 1 . . . . . . . 1 . . 2 . 1 . 1 . . . . . . . . .  . 1 .
#>                                                                                
#> HES4     . . . . . . . . . . . .  . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . 2 . . 1 . . . 9 1 . . 11 . . . 1 . . 5 . . . 1 . . . 2 . . . . . . .
#> AURKAIP1 . . 2 1 1 . . . . 1 . .  . 2 1 . 2 . . 2 . . 1 . 1 . 3 1 . . . 1 1 1 .
#> MRPL20   . . 1 . 1 . . . . . . .  . 1 . . 1 . . . . . . 2 . . 1 . . 1 . . . 1 .
#> SSU72    . . . . . 1 1 2 . . . .  . 1 . 1 . . 1 1 . 3 . 3 . . . 2 . . . 1 . . .
#> RER1     . . . . . 1 . 1 1 . . .  1 2 2 . . . 1 . . 1 1 1 . . 3 . . . . 2 1 . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . 3 1 . . 2 . . . . . 1 . . . . . . . . . . 1 3 . 2 . 1 2 1 . .
#> AURKAIP1 . 1 2 . 1 1 . 1 1 . . . . . 2 1 1 . . . 2 1 . 1 . . . . 3 . 3 1 . 1 .
#> MRPL20   . . . . 1 . . 1 . 1 . . 1 1 . . . . 2 . . 2 . . . . . 1 . . . . . 2 .
#> SSU72    . 1 1 . 1 1 . 1 1 . . . 2 . 2 2 1 . . . 2 1 1 . 1 . . . . 1 1 . 1 1 .
#> RER1     1 . 2 2 . 2 . . 3 1 . . 5 . 1 . . 1 . . 2 . 1 . 1 2 . . 1 1 . 1 . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . . . 1 1 . . . . . 1 . . . . . 1 . . . . . . . . . . . . . . .
#> AURKAIP1 1 . 2 . 2 1 2 1 . . . . . 1 . 1 . . 1 . . 1 . . 1 1 . . . . . . . . .
#> MRPL20   . . 1 . 1 . . . . 1 . 1 2 . 2 . . . . 1 1 1 . . . . . . 1 . . 1 1 . .
#> SSU72    . 1 . . . . . . . 1 . . 1 . . . . 1 . 1 . . 1 . 1 . . . . . . . . . .
#> RER1     . 1 . 2 . . 1 . 1 . 2 . 1 . . 1 . . . . . . 2 2 . . . . . . . . 2 . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . 1 . . . . . . . . . . . . . . . . . . . . 1 . . . . . 1 . . .
#> AURKAIP1 . . . . . . 1 . . . . . . . 1 . . . . 1 . . . . . . . 1 . . . . . . .
#> MRPL20   . . . . . . 1 . . 1 . . . 1 . . . . 1 . . . 1 . . . . 3 . . . . . . .
#> SSU72    . . . 1 . . . 1 . . 1 . . . . . . 1 . . . . . 1 . . . . . . . . . 1 .
#> RER1     . . . . . 1 . . . . . . . . . . . . . 1 . . 2 . . . . 1 . 1 . 1 . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    3 . . . . . . . . . . . 1 . . . . . . . . . . 1 . 2 1 . . . . . . 1 .
#> AURKAIP1 . . . . 1 . . . . . . . . . . . . . 1 . 1 . . . . . . . . . 1 1 . . .
#> MRPL20   2 . . . . 2 . . . . . . . 1 . . . . . . 1 . . . . 1 . . . 1 . . . . .
#> SSU72    2 2 . . . . . 1 . . . . . 1 . 1 . . . . . . . . . . . . . . . . . 1 .
#> RER1     . 1 . . . . . 1 . . . 1 . . . . 1 . 1 . . . 1 . . . . 1 . . . . . 1 1
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> AURKAIP1 . . . . . . . . . . . . . 1 . . 1 . . . 1 2 . . . . . . 2 . . . . . .
#> MRPL20   . . . . . . . 1 . . 1 1 . . . . . . 1 2 . . . . . . . . . . . . . . .
#> SSU72    . . . . 1 . . 1 . . . . . 1 . . . 1 . . . . . . . . . . . . . . . . .
#> RER1     . . . . . . . . . . . 1 1 . . . . . . . 1 1 . 1 . . . . . . . . . 1 .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . .
#> ISG15    . . 1 . . . . . . . . 1 . . . 1 . . 2 . . . . . . . . . . . . . . . .
#> AURKAIP1 . . . . . . . . . 1 . . 1 1 . . 1 . . . 1 . . . . . . . . . . . . 1 1
#> MRPL20   . . . . 1 . . . . . . . . . . . . . 1 . . . . 1 . . . . . 1 . . . . 1
#> SSU72    . . 1 . . . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . .
#> RER1     . 1 . 1 . 1 . 1 . . . . . 1 . . . . . . 1 3 . . . 1 . . . . . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . . . . . . . . . . . . . . . . . . . . 1 . . 2 1 . . 5 1 . . 1
#> AURKAIP1 1 1 . . . . . . . . . . . . . . . . 1 . . . . . 2 . . . . 1 1 . . . 2
#> MRPL20   . . . . . . . 1 . . 2 . 1 . . . 1 . . . . . . 1 . . . . . . . 1 . . .
#> SSU72    . . . . . . . . . . . . . . . 1 . . . . . . . . . 1 . . . . . . . 1 1
#> RER1     2 . . . . . . . . . . . . . . . . . . . . . . 1 2 . . . 1 1 . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    1 7 . . . . . 2 . . 1 . 1 . 1 . . . . . . 1 . . . 1 . . . . . 1 . . .
#> AURKAIP1 . . . 1 1 1 2 . . . . . . 1 2 1 1 . . . 1 1 . 3 . 1 1 . 1 . 1 1 . . .
#> MRPL20   . . . . 1 . . . . . . . 1 . 1 . . . . . . . . . . . . 1 2 . . 1 . . .
#> SSU72    1 1 . . . 1 . 1 . . . . . . 1 . . . . 2 . . . 1 1 1 . . . . 1 . . 1 .
#> RER1     . 1 . . . 1 . . 2 . . 1 . . . . . . . . . 1 . . . 1 . 1 2 . . . 1 2 1
#>                                                                               
#> HES4     . . . . . . . . . . . . . . 1 . . . . . . . . . . 1 . . . . . . . . .
#> ISG15    . . . . . . . . . 8 . 1 . . . 1 2 . . . . 1 2 . . . 1 2 1 . . . . 1 2
#> AURKAIP1 . . 1 . . 2 . 1 1 . . 1 . . . . . . . 1 1 . 1 . . 2 . . . 1 . 1 . . 1
#> MRPL20   . . . . . 1 . . . 1 . . . . . 1 1 . . . . 1 . 1 1 . 1 1 1 . . . . . .
#> SSU72    1 1 3 . . . . . . . . 1 . 1 1 1 . . . . . 1 1 . . . . . . 1 1 1 . 4 1
#> RER1     . . 1 . 2 . . 1 1 . . . . 1 . . . . 1 1 1 . . . . 1 . . . 2 1 1 1 . 1
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . . . . . . . 1 . . . . . . . . . . . . . . . . . . . . . . . . . .
#> AURKAIP1 . . . . . . . . . . . . 1 . . 1 1 1 . . 1 . 1 . . . . 1 . 1 . . 1 . .
#> MRPL20   . . . . . . 1 . . . . . . . . . . 1 . . 1 . . . . . . . . . . 3 . . .
#> SSU72    . 1 . 1 . . 1 . . . . 1 1 . . . 1 . . 1 1 . . . 1 1 . . . . . 1 . . 1
#> RER1     . . . 1 . . . . . . . . . 1 . . . 1 . . . . 1 . . . . . . . . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . . 2 . . 1 . . 1 . . . . . . . . 1 . . . . . . . . . 5 . . . . 3 . .
#> AURKAIP1 1 . 1 . . 1 . . . 1 1 . . . . 1 . . 1 1 . 1 2 . 1 2 2 . . 1 . 1 . . 1
#> MRPL20   . . . . . . . . . . . . 1 . . . . . . . . . . . 1 1 . . . . . . . . .
#> SSU72    . . . . . 1 1 . . 1 . . . . . . . . 1 . 1 . . . . 2 . 1 . 2 1 . . 1 .
#> RER1     1 . 1 . . . . . 3 . . . . 1 . 1 1 . . 1 2 . . 1 . . 1 1 . . 1 3 1 . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    1 . 1 . . . . . . . . . . . . 1 1 3 . . . . . . . . . . . 1 . . . . .
#> AURKAIP1 1 . 1 . 1 1 1 . 1 1 1 . 1 3 . . 2 . . . . 1 1 . . . . . 1 . 1 . . . .
#> MRPL20   1 . . . . 1 . 1 1 . . . . 1 . . 1 . . . . . . . . . . . 1 1 . . . . .
#> SSU72    . 1 . 2 . 1 . . . 1 1 . . . . . . . . . . 1 . . . . . . . 1 . . 2 . .
#> RER1     . 1 . . . 1 . . . 1 . . . . . 2 1 . 2 . 2 . . . . . . . 4 . 3 . . 2 1
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . 2 . . . 1 . . . . . 1 . . 1 1 . . . . . . . . . . 1 . 1 . . . . . .
#> AURKAIP1 . 1 . . . . . . 1 . 1 . 2 . . . . . . 2 . 1 . . . . 1 . 1 . . 1 1 1 .
#> MRPL20   1 . 1 . . 1 . . . 1 1 . . . . . . . . . . . . . . . . . . . . . . . 1
#> SSU72    . 1 . . . . . . 2 1 . 1 . . . 1 . . 1 . . 1 . . 2 . 1 . . 1 . 1 . . .
#> RER1     . 1 2 1 1 3 1 1 2 5 1 . 1 . . . . . 2 . . . . 2 . . . 1 . 2 . 2 . 1 .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . . . . . . . .
#> ISG15    6 . . . . 1 . . . . . . . 1 . . . . . 1 5 . . . 1 . . . . . . . . . .
#> AURKAIP1 . . . . 1 . . . 2 . . 1 . . . . . . . . . . 1 1 . 1 1 . . . . . 5 . .
#> MRPL20   . 2 . 1 . . . . . 1 . . . . . . . . . . 3 . . . . . . . . . . . . . .
#> SSU72    . 1 . . . . . . . . 1 2 . . . . . . 1 . 1 . . . . . . . . . . . . . .
#> RER1     1 . 1 . . 1 . . 1 . . 2 . . 1 . . . . . . 1 . . . . . . . . . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    1 . . . . . . . . . 2 . . . . . . . 1 . . . 4 . . . . . . . . . . . .
#> AURKAIP1 . . . . . . . . . . . . . . 1 . . . . . . 1 . . . . 1 . . . . . . . .
#> MRPL20   . . . . . . . . 1 . . . . . . . . . . 1 1 . . . . . . . . . . . . . .
#> SSU72    1 1 . . . . . . . 1 . . . . 1 . . . 1 . . 1 . . . . . . . 1 . . . . .
#> RER1     . . . . . . . . . . . . . . . 1 . . . . . . . . . . . . . . . . 1 . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . 1 . . . . .
#> ISG15    . . . . . . . . 1 . . 1 . . . . . . . . . . . 1 . . . . . . . . . 1 .
#> AURKAIP1 1 . . . . . . . 1 . . 1 . . . . 1 . . . . . . . . . . . . . 1 . . . .
#> MRPL20   . . . . . . . . 1 . . 2 . . . . . . . . . . . . . . . . . . 1 . 1 . .
#> SSU72    . . . . . 1 . . . . . 1 . . . . . . . . . . 1 . . . 1 1 . . 1 . . . .
#> RER1     . . . . . 1 1 . . . . . 1 . 1 . . 2 . . . . . 1 . . . . . . . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> ISG15    . 1 . . . 1 . . . 1 1 . . . . . . . . . 1 1 . . . . 1 . . . 1 . . . .
#> AURKAIP1 . . . . . . 1 . . . . . . . . . . 1 . 4 . . 1 . 2 . . . . 2 2 . . . .
#> MRPL20   . . . 1 . . . . . . . . 1 2 . 1 1 1 . 1 . 1 . . . . . . . 1 . 1 . . .
#> SSU72    . . . . . 1 . . . . . . . . 1 . . . . . . . . . . . 1 . . 1 . 2 1 . .
#> RER1     . 2 . . . . . . . . . . . . . 1 . 1 . . . . . 1 . . . . 1 . . . . . .
#>                                                                               
#> HES4     . . . . . . . . . . . 2 . . . .  . .  .  . .  . . . . .  . . .  . . .
#> ISG15    . . . . . . 2 . . . 1 . . 1 . 4 19 5 15 10 2 15 7 7 4 8 12 2 8 14 4 7
#> AURKAIP1 2 1 1 . . . . . . . 1 . . . . 1  . .  .  . 1  . . . . .  . 1 .  . 1 .
#> MRPL20   . . . . . . . . . . . . . . . .  . .  1  . .  . . . . .  . . 1  . . .
#> SSU72    2 . . 1 . . . . . . . . . . . .  1 .  .  . .  . . . . .  . 1 .  . . .
#> RER1     1 1 1 . . . 1 . 1 . . . . 1 1 1  1 .  .  . 1  . . . . .  . 1 .  . . .
#>                                                                               
#> HES4      . . . . .  . . . . . .  .  . . .  .  . . . . .  . 1 . . 1  . .  .  .
#> ISG15    28 5 7 8 6 16 7 5 7 6 4 27 12 9 . 19 16 7 9 4 5 10 8 4 1 5 17 6 20 11
#> AURKAIP1  1 . . . .  . . . . . .  .  1 . .  .  1 1 1 . .  1 2 . . .  1 .  1  .
#> MRPL20    . . . . .  . . . . . .  1  2 . .  .  . . . . .  2 1 . . .  . .  2  .
#> SSU72     . . 1 . .  . . . . . .  1  1 . .  .  . . . . .  . . . . 1  . .  .  1
#> RER1      . . . . .  . 1 . . . .  1  . . .  .  . . . . .  . . . . .  . .  .  .
#>                                                                               
#> HES4      . .  .  1 . . .  . . .  .  . 1  .  .  . .  .  . .  .  . . . . . .  .
#> ISG15    11 8 20 23 7 1 6 11 4 7 12 22 6 16 20 10 7 13 12 1 11 16 5 3 4 6 8 17
#> AURKAIP1  2 .  .  . . . .  . . .  1  . 1  .  .  . .  .  . .  1  . . . . 1 .  1
#> MRPL20    . .  .  . . . .  1 . .  .  1 .  .  .  . .  1  . .  .  . . . . . .  .
#> SSU72     . .  2  . . . 1  1 . 1  .  . 1  .  .  . .  .  1 .  .  . 1 . . . .  .
#> RER1      . .  .  . . 1 .  1 . .  .  . .  .  1  . .  1  . .  .  . . . . . .  .
#>                                                                                
#> HES4      .  . . .  .  . 1 . .  . . . .  . .  .  . . . . . . . . .  1  .   2  .
#> ISG15    13 17 3 7 61 17 9 1 9 10 . 5 7 18 7 20 10 3 8 4 4 7 5 5 9 45 68 155 79
#> AURKAIP1  .  1 . .  1  1 1 . .  . . . .  . .  .  . 1 1 . . 1 . . .  .  1   .  2
#> MRPL20    .  . . .  2  . 1 . .  . . . 1  . .  .  . . . . . . . . .  .  .   .  .
#> SSU72     .  1 1 .  .  . . 1 1  1 1 . 1  . .  .  . . . . . . . . .  .  .   .  .
#> RER1      .  . 1 .  .  1 1 . 1  . . . .  2 .  .  . . . . . 2 . . 1  .  .   .  1
#>                                                                                
#> HES4       1  .  .  .  .  .  .   .  .   1  .   .   2   1  .   .  .   .  .  .  .
#> ISG15    220 89 75 86 91 74 99 106 21 107 45 109 141 168 53 137 29 142 49 96 84
#> AURKAIP1   1  1  .  .  .  .  .   1  .   1  .   .   1   .  2   .  1   1  .  .  .
#> MRPL20     .  .  .  1  .  .  .   .  .   .  .   1   .   .  .   .  1   .  .  .  1
#> SSU72      1  .  .  .  .  .  .   .  1   .  .   .   .   .  .   .  .   .  .  .  .
#> RER1       .  .  .  .  .  .  2   .  .   2  .   .   1   .  .   1  1   .  .  .  1
#>                                                                              
#> HES4      2  .  .   2  .  .   .  .  .  .  .  .  .   6  .  .   4  .  .   .   .
#> ISG15    54 79 95 182 23 61 103 71 23 91 16 68 62 217 63 50 148 65 31 132 101
#> AURKAIP1  .  .  1   .  1  .   .  1  .  .  .  1  .   1  1  .   1  .  .   .   .
#> MRPL20    .  .  .   .  .  .   .  .  .  .  .  .  .   1  .  .   .  .  .   .   .
#> SSU72     .  .  .   1  .  .   .  1  1  .  1  .  .   .  .  .   .  .  .   .   1
#> RER1      .  2  1   1  .  .   .  1  .  1  .  .  .   1  .  .   .  .  .   1   .
#>                                                                              
#> HES4      1  .   2  .  .  1   1  .  1   .  1   3  .  .   .   .  .  .  .  2  .
#> ISG15    75 67 173 70 37 43 145 65 72 101 56 125 44 77 165 114 90 80 42 86 99
#> AURKAIP1  .  1   .  .  .  .   .  .  .   .  .   2  .  .   1   .  1  .  1  1  .
#> MRPL20    .  .   .  .  .  .   .  .  .   .  .   2  .  .   .   .  .  1  .  .  .
#> SSU72     .  .   .  .  .  .   .  .  1   .  .   .  .  .   .   .  .  .  .  .  2
#> RER1      .  1   .  .  .  .   .  .  .   1  .   1  .  .   .   1  .  1  .  .  1
#>                                                                                
#> HES4       1  .   1   2  .  .  1  .  .   5   .   .  .  .   .   .   .   .   2  .
#> ISG15    215 51 202 187 22 65 31 70 56 195 121 110 64 11 168 113 102 125 159 77
#> AURKAIP1   .  .   1   .  .  .  1  .  .   .   1   .  .  .   .   .   1   .   .  1
#> MRPL20     .  .   .   .  1  .  .  .  .   .   .   .  .  .   .   .   .   1   .  .
#> SSU72      .  .   .   .  .  .  .  .  .   1   .   1  .  .   .   .   .   .   .  .
#> RER1       1  .   1   .  1  .  1  .  .   .   .   1  .  .   1   2   1   2   .  3
#>                                                                                
#> HES4       .  .   .  .   .  2  .  .  .  .  .  .   1 . . .  . . .  . .  . . .  .
#> ISG15    190 33 137 30 142 42 89 61 25 54 58 73 114 4 8 8 20 . 6 21 4 10 4 1 10
#> AURKAIP1   .  .   2  .   .  1  3  2  1  .  .  .   . . . .  . 1 .  1 1  1 . .  .
#> MRPL20     .  1   .  .   .  .  .  .  .  .  .  .   . . 1 1  1 . .  1 1  . . .  .
#> SSU72      .  .   .  .   .  .  .  1  .  .  .  .   1 . . 1  . . .  . .  1 . .  .
#> RER1       1  1   .  .   2  .  1  2  .  1  1  .   . . . .  . . .  . .  . . .  1
#>                                                                              
#> HES4      . . . . .  . . .  . . . .  . . . . . . . . .  . .  1  . .  . .  . .
#> ISG15    10 9 4 4 2 24 7 6 15 8 4 1 10 4 3 9 6 5 2 . 2 12 6 19 13 3 17 3 12 3
#> AURKAIP1  . . . . .  . . 1  1 . . .  . . . . . . . . .  . .  .  . 1  . .  . .
#> MRPL20    . . . . 1  . . .  . . . 1  . . . . 1 1 . . .  . 1  .  . .  . .  . .
#> SSU72     . . . . 2  . . .  . 1 . .  . . . . . . . . .  1 .  .  . .  . .  2 .
#> RER1      1 . 1 . .  . . .  . . . 1  . 1 . . . . . . 1  1 1  .  1 .  . .  . .
#>                                                                                
#> HES4      . . .  . . . . . .  . . . . . .  . . .  . . . . .  . . . .  . . . . .
#> ISG15    10 5 5 11 . 9 8 6 6 18 8 5 2 4 1 18 1 2 35 5 6 5 5 19 6 8 7 10 3 8 8 6
#> AURKAIP1  . . .  . . . . 1 .  . . . . . .  . . .  1 . . 1 .  . . 1 .  . . 2 . .
#> MRPL20    . . .  . . . . 1 .  . . 1 . . .  . . .  . . . 1 .  1 . . 1  . . . . .
#> SSU72     . . .  1 . . . . .  . . . . . .  1 . .  . . . . 1  . . . .  . . 1 . .
#> RER1      1 . .  . . . . 1 .  . . . 1 . .  . . .  . . . . .  . . . .  . . . 1 .
#>                                                                                
#> HES4     .  . .  .  . . . .  . .  . . . . . . . .  1 . . .  .  . . . .  . . . .
#> ISG15    1 10 6 13 14 2 8 8 14 3 39 4 7 1 9 9 4 8 61 2 5 5 19 12 6 6 2 12 9 2 1
#> AURKAIP1 .  . .  .  1 . . .  1 1  . . . . . . . 1  . 1 . .  .  . 1 . .  . . . .
#> MRPL20   .  1 .  .  1 . . .  1 1  . . . . . . . .  1 . . 1  1  . . 1 .  . . . .
#> SSU72    .  . .  .  . . . .  . .  . . . . . . . .  . . . .  .  . 1 . .  . . . .
#> RER1     .  . .  .  2 . . .  . .  . . . . . 1 . .  . . 1 .  .  . . . 1  . . 1 .
#>                                                                               
#> HES4     . .  . . .  . . . . . . . . . . . . . . . . . .  .  . . . 1 . . . . .
#> ISG15    3 9 11 2 5 18 1 7 1 1 6 . 6 2 4 2 5 9 7 9 6 4 7 18 15 1 1 3 7 6 4 1 3
#> AURKAIP1 . .  1 . .  1 . . . . . . . 1 . . . . . . . . .  .  . . . . . . 1 . .
#> MRPL20   . .  . . .  1 . . . . . . . . . . . . . . . . .  .  . . . . . . . 1 .
#> SSU72    . .  1 . .  . . . . . . . . . . . . . . . . 1 .  .  . . . 1 . . . . .
#> RER1     . .  1 . .  . 1 1 . . . . . . 1 . . . . 1 . . .  1  . . . . . 1 . . 1
#>                                                                                
#> HES4     . . .  .  . . . .  . .  . . . 1 . . . . .  . . . .  . . . . . . . .  .
#> ISG15    6 2 8 16 16 1 5 2 31 2 15 1 3 5 2 9 7 9 3 11 1 1 5 10 3 7 7 6 8 8 1 17
#> AURKAIP1 . . .  .  . . . .  . .  1 . . . . . 1 . .  2 1 . 1  2 . . 1 . 1 . .  .
#> MRPL20   1 . .  .  . . . .  . .  . . . . 1 . . . .  . . . .  . . . . . 1 . .  .
#> SSU72    . . .  .  . . . .  1 .  1 2 . . . 1 . . .  . . . .  . . . . 1 . . .  .
#> RER1     . 1 .  .  . . . .  . .  . . . . . . . . .  . . . .  . . . . . . . .  .
#>                                                                               
#> HES4     . . . . . .  . 1 1 . 1 . . .  .  . . .  . .  . . 2 .  . . .  . . .  1
#> ISG15    4 1 7 3 9 4 10 3 3 4 3 5 8 . 10 12 3 3 10 4 19 7 6 3 10 4 1 13 5 . 50
#> AURKAIP1 2 . . . . .  . 3 2 . . . . .  .  . . .  . .  . . . .  . . .  . . .  .
#> MRPL20   . . . . 1 .  . . . . . . . .  .  . . .  . .  . . . .  . . .  . 1 .  1
#> SSU72    . . . . . 1  . . . . . . 1 .  .  . . .  1 .  . . . .  1 . .  . . 1  .
#> RER1     1 . 1 . . .  . . . . 1 . . .  .  . . .  . .  . . . .  . . .  . . .  1
#>                                                                               
#> HES4      2  1  4  .  1  .  .  .  .  .  .   1  3   1   .  1  2  2  .  .   8  .
#> ISG15    93 76 67 11 71 64 57 36 85 11 64 129 27 149 151 81 73 51 80 58 189 20
#> AURKAIP1  .  .  .  .  2  .  .  .  .  .  .   1  .   1   .  1  .  2  .  .   .  .
#> MRPL20    .  .  1  .  .  .  .  .  .  .  .   .  .   .   .  .  .  .  .  1   .  .
#> SSU72     .  1  .  1  .  .  .  .  .  .  .   1  .   .   .  .  1  1  .  .   1  .
#> RER1      .  .  .  1  .  .  1  .  .  .  .   .  1   3   .  .  1  1  2  1   1  .
#>                                                                              
#> HES4      .  2   2  1  .  4  1   1  .  .   .  .  .  1  .  .  1   5   2   2  .
#> ISG15    57 76 169 75 32 33 46 115 73 44 125 94 85 60 64 66 42 169 104 126 97
#> AURKAIP1  .  .   .  .  2  .  .   .  .  .   2  .  .  .  .  1  1   1   1   1  1
#> MRPL20    .  .   .  .  .  .  .   .  .  1   .  .  .  .  1  .  .   .   .   .  .
#> SSU72     .  .   1  .  .  .  1   .  2  .   3  1  .  1  .  .  .   .   1   .  .
#> RER1      .  .   1  2  1  .  1   1  .  1   1  .  .  .  .  .  .   1   .   1  .
#>                                                                              
#> HES4      .  .  .  .  .  .  2   2 .  2  1  . .  .  .  .  .  .  .  .   1  .  2
#> ISG15    15 57 35 67 38 61 90 125 7 69 78 22 5 42 49 85 45 32 43 96 156 66 53
#> AURKAIP1  .  2  .  .  .  2  1   . .  .  1  . .  .  .  .  1  1  .  .   .  1  .
#> MRPL20    .  1  .  .  .  .  .   . 1  .  .  . .  .  1  2  1  .  .  1   .  .  .
#> SSU72     .  .  .  .  .  .  .   1 .  .  1  . .  1  1  1  .  .  .  1   .  .  .
#> RER1      .  .  .  .  2  .  1   3 .  .  3  . .  2  1  .  .  .  .  .   1  .  1
#>                                                                              
#> HES4      1  .   2   .  5   2  2  .  .  .  .  .  1  .  .  .  .  1   .   .   2
#> ISG15    55 66 119 117 54 199 33 85 64 49 70 31 65 24 32 40 67 66 105 101 167
#> AURKAIP1  .  .   1   1  .   .  .  1  1  .  .  .  .  .  .  .  .  1   1   2   .
#> MRPL20    .  .   1   .  .   .  .  .  .  1  .  .  .  .  .  .  1  1   .   .   .
#> SSU72     1  1   .   .  1   .  .  .  .  .  .  .  1  .  .  1  3  .   .   1   3
#> RER1      .  .   .   2  1   1  .  .  1  .  .  1  .  1  .  1  .  .   1   1   .
#>                                                                                
#> HES4      .  1  .  1  1  1  1  .  .  .  .  .  . . . . .  .  .  . 1 . .  1 . . .
#> ISG15    15 67 43 20 76 88 60 32 42 35 75 68 10 1 8 2 4 18 12 11 9 2 6 13 5 7 3
#> AURKAIP1  .  1  .  .  1  .  .  .  .  .  2  .  . . . . 1  .  .  . . . 1  . 1 . .
#> MRPL20    .  .  .  .  .  .  .  .  1  .  1  .  . . . . .  .  .  . . . .  . . . .
#> SSU72     .  .  .  .  .  .  .  .  2  .  .  .  1 1 1 . .  1  .  1 1 1 1  . 1 . 2
#> RER1      .  .  .  .  1  .  .  .  .  .  1  1  . . . . 1  .  1  . . 1 .  2 . . .
#>                                                                               
#> HES4     . .  . .  .  . . . .  .  . . . .  .  . . . .  . .  .  . .  .  .  .  .
#> ISG15    5 2 10 7 15 11 9 7 4 12 12 8 9 1 10 13 2 9 8 23 3 17 20 7 12 13 27 20
#> AURKAIP1 . 1  . .  .  . . . .  .  . 1 1 .  .  . . . .  . .  .  . .  .  1  .  2
#> MRPL20   . .  . 1  1  . . . .  .  . . . .  .  . . . .  . 1  .  . .  .  .  .  1
#> SSU72    1 .  . .  .  . . . .  .  . 1 . .  .  . . . .  . .  1  2 .  2  .  1  .
#> RER1     1 .  . .  .  . . . .  .  . . . .  .  . 1 . .  . .  .  . .  1  .  .  1
#>                                                                               
#> HES4     . .  . . .  . . . . .  1  .   6   4   4   3   .   1   2  1  .   3   3
#> ISG15    5 8 11 5 8 43 8 6 3 5 33 56 215 138 154 181 121 182 356 63 81 143 115
#> AURKAIP1 1 .  . 2 .  . . . . .  .  .   .   1   1   .   .   .   .  .  .   1   .
#> MRPL20   . .  1 . 1  . . . . .  .  .   .   .   .   .   2   .   .  .  .   .   .
#> SSU72    . .  . . .  . . . . .  .  .   1   .   .   .   .   .   .  .  .   .   .
#> RER1     . .  . 1 .  1 . . 1 .  1  1   3   .   .   1   .   1   .  .  .   .   .
#>                                                                               
#> HES4       .  .   .  .   2  2   1   .   .   3  .  .   1   .  .   .   3   3   .
#> ISG15    119 94 123 69 219 31 193 260 137 104 65 31 120 114 85 153 215 187 125
#> AURKAIP1   .  .   .  .   1  .   .   .   .   .  .  1   .   1  .   .   .   .   1
#> MRPL20     .  .   .  .   .  .   .   .   .   .  .  .   .   .  .   .   .   .   .
#> SSU72      .  .   .  .   .  .   .   .   .   .  .  .   .   .  .   .   1   1   .
#> RER1       .  1   .  .   3  .   .   .   1   1  .  .   .   .  .   .   .   .   1
#>                                                                                
#> HES4      .   2   4   1   .   .   .   1   .   1   .  .   4  .   2  .  .   2   1
#> ISG15    14 143 189 141 170 169 132 210 242 118 300 98 216 24 128 42 32 124 243
#> AURKAIP1  .   .   .   .   1   .   2   1   1   .   1  .   1  .   .  .  .   1   .
#> MRPL20    .   .   .   .   .   .   1   .   .   .   .  .   .  .   .  .  .   .   .
#> SSU72     .   .   .   .   .   .   .   1   .   .   .  1   .  .   .  .  .   1   .
#> RER1      .   .   .   2   .   .   1   .   .   .   .  .   .  .   1  .  1   1   .
#>                                                                               
#> HES4      1   .   1   2   3   2  .  .   .   2   .   1   1   2  .   5  .  1   2
#> ISG15    43 110 176 170 145 328 49 83 163 100 135 130 167 200 89 297 53 67 212
#> AURKAIP1  .   .   .   .   1   .  .  .   .   1   .   .   1   .  .   .  1  1   .
#> MRPL20    .   .   .   .   .   .  .  .   .   .   .   .   .   .  .   .  .  .   .
#> SSU72     .   1   .   .   1   .  .  1   1   .   .   1   .   .  .   .  .  3   .
#> RER1      .   .   .   1   .   .  .  1   .   .   2   .   .   .  .   1  1  1   .
#>                                                                               
#> HES4      .  .   3   4  .   2   2   4   4  10   5  2  3   6   4   .   .  1   .
#> ISG15    60 99 222 120 70 219 262 144 226 338 383 82 89 214 264 171 102 69 144
#> AURKAIP1  .  1   .   1  .   .   .   .   .   .   .  .  .   .   .   .   .  .   .
#> MRPL20    .  .   .   .  .   .   1   .   .   1   .  .  .   .   .   .   .  .   .
#> SSU72     .  .   .   .  .   .   .   .   .   .   .  .  .   .   .   .   .  .   .
#> RER1      1  .   2   1  .   1   1   .   1   .   2  1  .   .   .   .   .  1   1
#>                                                                                
#> HES4       2   4   5   2   .  1   4   6   1   6   .   3  . . . . .  . . . . . .
#> ISG15    101 191 169 101 120 36 171 260 341 211 191 182 11 4 6 3 3 11 5 8 3 5 3
#> AURKAIP1   1   1   .   .   .  .   .   1   1   .   .   .  . . . . .  1 . . . . .
#> MRPL20     .   .   .   .   .  .   .   .   .   .   .   1  . . . . .  . 1 1 1 . .
#> SSU72      .   .   .   .   .  .   1   .   .   .   .   1  . 1 . . .  . . . 1 . .
#> RER1       .   .   .   1   .  .   1   1   .   .   .   .  1 . . . .  . . . . . 1
#>                                                                              
#> HES4      . . . .  . .  .  . .  . . . . . . . .  . . . . . . . .  . .  . .  .
#> ISG15    14 3 3 9 26 . 11 10 5 12 9 6 . 2 1 1 9 12 6 2 6 4 3 4 3 16 3 13 4 21
#> AURKAIP1  . . . .  . .  .  . .  . . . . . . . .  . . . . . . . .  . .  . 1  .
#> MRPL20    . . . .  1 .  .  1 .  . . 1 . . . . .  . . . . . . . .  . .  1 .  .
#> SSU72     1 . . .  . .  .  . .  . . . . . 1 . .  . . 1 . . . . .  . .  . .  2
#> RER1      . . . .  . .  .  1 1  . . . . . . . .  . . . 1 . . . .  . .  . .  1
#>                                                                                
#> HES4      . . . . . . . . . . . .  . . . .  .  .  . . . . . . . . . . . . . . 1
#> ISG15    25 3 7 2 2 4 4 1 6 4 6 3 28 3 2 7 10 16 36 8 5 1 1 1 3 . 4 5 7 6 5 9 8
#> AURKAIP1  . . . 1 . . 1 . . 1 . 1  . . . .  .  2  . . . . . . . . . . 1 . . . .
#> MRPL20    . . . . . . . . . . . .  2 1 . .  .  .  . . . . . . . . . 3 . . . 1 1
#> SSU72     . . . . . . . . . . 1 1  . 1 1 .  .  .  1 . . . 1 . . . . . . 1 . . .
#> RER1      . . . 1 . . . . . 1 . .  . 1 . .  .  .  1 2 . . . . . 1 1 . . . 1 . .
#>                                                                                
#> HES4     .  . . . . . . . . . . . . . . .  .  . . . . . . . . . . . . . . . . .
#> ISG15    4 17 6 3 8 9 5 . 7 9 5 3 4 7 3 7 17 14 5 5 6 7 6 4 8 8 2 7 8 5 7 1 . 6
#> AURKAIP1 .  . . . . . . . . . . . 1 . 1 .  .  . . . . . . . 1 . . . . . . . . .
#> MRPL20   .  . . . . . . 1 . . . . . . . .  .  . . . . 2 . 2 . . . . . . . . . .
#> SSU72    .  . . . . . . 1 . 1 . . . . 1 .  .  . . . . . 1 . . . . 1 . . . . . .
#> RER1     .  . 1 2 1 . . . . . . . . . . .  .  . . 1 . . . . . . . . . 1 . . 1 .
#>                                                                               
#> HES4     .  . .  . . .  .  1  .  3  .   1  5  . .   1  1  1  5  .  1  .   .  .
#> ISG15    9 13 5 10 4 4 25 63 57 93 62 105 87 69 4 171 93 96 79 79 35 50 157 21
#> AURKAIP1 .  . .  1 . .  1  1  1  .  .   .  .  . .   1  .  .  1  .  .  .   .  .
#> MRPL20   .  . 1  . 1 .  .  .  .  1  .   .  .  . .   .  .  .  .  .  .  .   .  .
#> SSU72    1  . 1  . 1 .  .  .  .  1  .   .  .  . .   .  .  .  .  .  .  .   .  .
#> RER1     .  . .  . . .  .  .  .  .  1   .  1  . .   .  .  .  1  .  .  .   .  .
#>                                                                          
#> HES4      1  .  .   1  .  2  .   2  .   4  .  .  .  .  1  1  1   .  9   .
#> ISG15    80 54 64 173 64 75 72 113 81 106 39 83 75 39 36 76 56 153 51 123
#> AURKAIP1  .  .  .   .  .  .  .   .  1   .  .  .  .  2  .  .  1   .  .   1
#> MRPL20    .  .  .   .  .  .  .   .  .   .  .  .  .  .  .  .  .   .  .   1
#> SSU72     .  .  .   1  .  .  .   .  .   .  .  .  .  .  .  .  .   .  .   1
#> RER1      .  .  .   1  1  .  .   .  .   1  .  .  1  .  .  .  1   .  .   .

# scaled CPM
cpm <- edgeR::cpm(assay(example_sce))
assays(example_sce)$cpm <- cpm
pb <- aggregateToPseudoBulk(example_sce, assay = "cpm", scale = TRUE, verbose=FALSE)
head(assay(pb))

#>             ctrl101  ctrl107    stim101     stim107
#> HES4       49.61317   0.0000   511.1207    69.06912
#> ISG15     464.89019 474.7922 57690.0813 48047.65032
#> AURKAIP1 2174.67064 527.1645  2572.3399   616.75874
#> MRPL20    800.69487 463.4986   671.7521   612.32657
#> SSU72    2640.12114 873.4110  2443.2336   834.68816
#> RER1     1225.83834 636.7921  1196.8081   980.45076

# aggregate by cluster only
pb <- aggregateToPseudoBulk(example_sce, by = "cluster_id", verbose=FALSE)
length(assays(pb)) # single assay

#> [1] 1

head(assay(pb))    # n_genes x n_clusters

#>          B cells CD14+ Monocytes CD4 T cells CD8 T cells FCGR3A+ Monocytes
#> HES4           9             216           5           8               125
#> ISG15       1538           24271        1567         729              9999
#> AURKAIP1      74             175          64          40               102
#> MRPL20        41              71          81          27                42
#> SSU72         83             112          67          23                80
#> RER1          70             214          72          37               100