Differential expression for each assay

Perform differential expression for each assay using linear (mixed) models

dreamlet(
  x,
  formula,
  data = colData(x),
  assays = assayNames(x),
  contrasts = NULL,
  min.cells = 10,
  isCounts = TRUE,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  use.eBayes = TRUE,
  ...
)

# S4 method for dreamletProcessedData
dreamlet(
  x,
  formula,
  data = colData(x),
  assays = assayNames(x),
  contrasts = NULL,
  min.cells = 10,
  isCounts = TRUE,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  use.eBayes = TRUE,
  ...
)

Arguments

x

SingleCellExperiment or dreamletProcessedData object

formula

regression formula for differential expression analysis

data

metadata used in regression formula

assays

array of assay names to include in analysis. Defaults to assayNames(x)

contrasts

character vector specifying contrasts specifying linear combinations of fixed effects to test. This is fed into makeContrastsDream( formula, data, contrasts=contrasts)

min.cells

minimum number of observed cells for a sample to be included in the analysis

isCounts

logical, indicating if data is raw counts

robust

logical, use eBayes method that is robust to outlier genes

quiet

show messages

BPPARAM

parameters for parallel evaluation

use.eBayes

should eBayes be used on result? (defualt: TRUE)

...

other arguments passed to dream

Value

Object of class dreamletResult storing results for each cell type

Details

Fit linear (mixed) model on each cell type separately. For advanced use of contrasts see variancePartition::makeContrastsDream() and vignette https://gabrielhoffman.github.io/variancePartition/articles/dream.html#advanced-hypothesis-testing-1.

See also

Examples

 
library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)
#>   B cells...
#> 0.017 secs
#>   CD14+ Monocytes...
#> 0.017 secs
#>   CD4 T cells...
#> 0.018 secs
#>   CD8 T cells...
#> 0.016 secs
#>   FCGR3A+ Monocytes...
#> 0.021 secs

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)
#>   B cells...
#> 0.037 secs
#>   CD14+ Monocytes...
#> 0.056 secs
#>   CD4 T cells...
#> 0.044 secs
#>   CD8 T cells...
#> 0.022 secs
#>   FCGR3A+ Monocytes...
#> 0.055 secs
#> 

# show coefficients estimated for each cell type
coefNames(res.dl)
#> [1] "(Intercept)"  "group_idstim"

# extract results using limma-style syntax
# combines all cell types together
# adj.P.Val gives study-wide FDR
topTable(res.dl, coef="group_idstim", number=3)
#> DataFrame with 3 rows and 8 columns
#>         assay          ID     logFC   AveExpr         t     P.Value   adj.P.Val
#>   <character> <character> <numeric> <numeric> <numeric>   <numeric>   <numeric>
#> 1     B cells       ISG20   3.59771   11.5794   15.5721 1.01377e-18 3.55733e-15
#> 2 CD4 T cells       ISG15   5.50055   10.1672   21.5266 1.38871e-15 2.43649e-12
#> 3 CD4 T cells       ISG20   3.07310   10.7918   17.0119 1.37235e-13 1.60520e-10
#>           B
#>   <numeric>
#> 1   32.2930
#> 2   24.9379
#> 3   21.2155