Differential expression for each assay

Perform differential expression for each assay using linear (mixed) models

Usage

dreamlet(
  x,
  formula,
  data = colData(x),
  assays = assayNames(x),
  contrasts = NULL,
  min.cells = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  use.eBayes = TRUE,
  ...
)

# S4 method for dreamletProcessedData
dreamlet(
  x,
  formula,
  data = colData(x),
  assays = assayNames(x),
  contrasts = NULL,
  min.cells = 10,
  robust = FALSE,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  use.eBayes = TRUE,
  ...
)

Arguments

x

SingleCellExperiment or dreamletProcessedData object

formula

regression formula for differential expression analysis

data

metadata used in regression formula

assays

array of assay names to include in analysis. Defaults to assayNames(x)

contrasts

character vector specifying contrasts specifying linear combinations of fixed effects to test. This is fed into makeContrastsDream( formula, data, contrasts=contrasts)

min.cells

minimum number of observed cells for a sample to be included in the analysis

robust

logical, use eBayes method that is robust to outlier genes

quiet

show messages

BPPARAM

parameters for parallel evaluation

use.eBayes

should eBayes be used on result? (defualt: TRUE)

...

other arguments passed to dream

Value

Object of class dreamletResult storing results for each cell type

Details

Fit linear (mixed) model on each cell type separately. For advanced use of contrasts see variancePartition::makeContrastsDream() and vignette https://gabrielhoffman.github.io/variancePartition/articles/dream.html#advanced-hypothesis-testing-1.

See also

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)
#>   B cells...
#> 0.027 secs
#>   CD14+ Monocytes...
#> 0.026 secs
#>   CD4 T cells...
#> 0.031 secs
#>   CD8 T cells...
#> 0.027 secs
#>   FCGR3A+ Monocytes...
#> 0.032 secs

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)
#>   B cells...
#> 0.1 secs
#>   CD14+ Monocytes...
#> 0.15 secs
#>   CD4 T cells...
#> 0.11 secs
#>   CD8 T cells...
#> 0.065 secs
#>   FCGR3A+ Monocytes...
#> 0.13 secs
#> 

# show coefficients estimated for each cell type
coefNames(res.dl)
#> [1] "(Intercept)"  "group_idstim"

# extract results using limma-style syntax
# combines all cell types together
# adj.P.Val gives study-wide FDR
topTable(res.dl, coef="group_idstim", number=3)
#> DataFrame with 3 rows and 9 columns
#>         assay          ID     logFC   AveExpr         t     P.Value   adj.P.Val
#>   <character> <character> <numeric> <numeric> <numeric>   <numeric>   <numeric>
#> 1     B cells       ISG20   3.68911   11.5794   19.5111 5.61596e-23 1.97064e-19
#> 2 CD8 T cells       ISG20   3.73855   10.9147   22.5435 6.52412e-22 1.14466e-18
#> 3 CD4 T cells       ISG15   5.50055   10.1672   21.9010 2.62269e-15 3.06767e-12
#>           B     z.std
#>   <numeric> <numeric>
#> 1   41.8741   19.5111
#> 2   39.8860   22.5435
#> 3   24.3335   21.9010