Test if coefficient is different from a specified value

Test if coefficient is different from a specified value

Usage

# S4 method for dreamletResult
getTreat(fit, lfc = log2(1.2), coef = NULL, number = 10, sort.by = "p")

Arguments

fit

dreamletResult object

lfc

a minimum log2-fold-change below which changes not considered scientifically meaningful

coef

which coefficient to test

number

number of genes to return

sort.by

column to sort by

Value

DataFrame storing hypothesis test for each gene and cell type

See also

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce,
  assay = "counts",
  cluster_id = "cluster_id",
  sample_id = "sample_id",
  verbose = FALSE
)

# voom-style normalization
res.proc <- processAssays(pb, ~group_id)
#>   B cells...
#> 0.26 secs
#>   CD14+ Monocytes...
#> 0.37 secs
#>   CD4 T cells...
#> 0.28 secs
#>   CD8 T cells...
#> 0.13 secs
#>   FCGR3A+ Monocytes...
#> 0.39 secs

# Differential expression analysis within each assay,
# evaluated on the voom normalized data
res.dl <- dreamlet(res.proc, ~group_id)
#>   B cells...
#> 0.16 secs
#>   CD14+ Monocytes...
#> 0.34 secs
#>   CD4 T cells...
#> 0.2 secs
#>   CD8 T cells...
#> 0.12 secs
#>   FCGR3A+ Monocytes...
#> 0.21 secs

# show coefficients estimated for each cell type
coefNames(res.dl)
#> [1] "(Intercept)"  "group_idstim"

# extract results using limma-style syntax
# combines all cell types together
# adj.P.Val gives study-wide FDR
getTreat(res.dl, coef = "group_idstim", number = 3)
#> DataFrame with 3 rows and 7 columns
#>               assay     logFC   AveExpr         t     P.Value   adj.P.Val
#>         <character> <numeric> <numeric> <numeric>   <numeric>   <numeric>
#> 1 FCGR3A+ Monocytes   5.28241   11.9132   28.4893 9.58913e-23 3.63291e-19
#> 2 FCGR3A+ Monocytes   7.18747   11.1852   28.0168 1.61786e-22 3.63291e-19
#> 3 FCGR3A+ Monocytes   5.06992   10.4846   22.0644 1.01370e-19 1.51751e-16
#>           B
#>   <numeric>
#> 1   42.7133
#> 2   41.5764
#> 3   35.8064