Processing SingleCellExperiment to dreamletProcessedData

For raw counts, estimate precision weights using linear mixed model weighting by number of cells observed for each sample. For normalized data, only weight by number of cells

Usage

processAssays(
  sceObj,
  formula,
  assays = assayNames(sceObj),
  min.cells = 10,
  min.count = 10,
  min.samples = 4,
  isCounts = TRUE,
  normalize.method = "TMM",
  useCountsWeights = TRUE,
  pmetadata = NULL,
  pkeys = NULL,
  quiet = FALSE,
  BPPARAM = SerialParam(),
  ...
)

Arguments

sceObj

SingleCellExperiment object

formula

regression formula for differential expression analysis

assays

array of assay names to include in analysis. Defaults to assayNames(sceObj)

min.cells

minimum number of observed cells for a sample to be included in the analysis

min.count

minimum number of reads for a gene to be considered expressed in a sample. Passed to edgeR::filterByExpr

min.samples

minimum number of samples passing cutoffs for cell cluster to be retained

isCounts

logical, indicating if data is raw counts

normalize.method

normalization method to be used by calcNormFactors

useCountsWeights

use cell count weights

pmetadata

sample-specific metadata the varies across cell types. This is merged with colData(sceObj) for each assay to make variables accessable to the formula

pkeys

array of two strings indicating sample identifier and cell type identifier columns in pmetadata

quiet

show messages

BPPARAM

parameters for parallel evaluation

...

other arguments passed to dream

Value

Object of class dreamletProcessedData storing voom-style normalized expression data

Examples

library(muscat)
library(SingleCellExperiment)

data(example_sce)

# create pseudobulk for each sample and cell cluster
pb <- aggregateToPseudoBulk(example_sce, 
   assay = "counts",    
   cluster_id = 'cluster_id', 
   sample_id = 'sample_id',
   verbose=FALSE)

# voom-style normalization
res.proc = processAssays( pb, ~ group_id)
#>   B cells...
#> 0.018 secs
#>   CD14+ Monocytes...
#> 0.018 secs
#>   CD4 T cells...
#> 0.022 secs
#>   CD8 T cells...
#> 0.017 secs
#>   FCGR3A+ Monocytes...
#> 0.02 secs

# Differential expression analysis within each assay,
# evaluated on the voom normalized data 
res.dl = dreamlet( res.proc, ~ group_id)
#>   B cells...
#> 0.039 secs
#>   CD14+ Monocytes...
#> 0.061 secs
#>   CD4 T cells...
#> 0.047 secs
#>   CD8 T cells...
#> 0.026 secs
#>   FCGR3A+ Monocytes...
#> 0.052 secs
#>