Processing expression data from assay

For raw counts, estimate precision weights using linear mixed model weighting by number of cells observed for each sample. For normalized data, only weight by number of cells

Usage

processOneAssay(
  y,
  formula,
  data,
  n.cells,
  min.cells = 10,
  min.count = 10,
  min.samples = 4,
  isCounts = TRUE,
  normalize.method = "TMM",
  useCountsWeights = TRUE,
  quiet = TRUE,
  BPPARAM = SerialParam(),
  ...
)

Arguments

y

matrix of counts or log2 CPM

formula

regression formula for differential expression analysis

data

metadata used in regression formula

n.cells

array of cell count for each sample

min.cells

minimum number of observed cells for a sample to be included in the analysis

min.count

minimum number of reads for a gene to be considered expressed in a sample. Passed to edgeR::filterByExpr

min.samples

minimum number of samples passing cutoffs for cell cluster to be retained

isCounts

logical, indicating if data is raw counts

normalize.method

normalization method to be used by calcNormFactors

useCountsWeights

use cell count weights

quiet

show messages

BPPARAM

parameters for parallel evaluation

...

other arguments passed to dream

Value

EList object storing log2 CPM and precision weights